Move a paragraph between sections
First sentence of "Downstream patch generation process"
talks about `upstream_ref` like we already knew what it is.

Then the last paragraph of
"Downstream patches to be applied in the spec file"
talks about it as well (IIUIC), but this time explains it.

So I'd suggest moving that paragraph
('The important part of the patch generation ...')
to the beginning of "Downstream patch generation process".